Fix packaging issues

+++ package_linux_amd64.sh
@@ -8,6 +8,7 @@ echo "# Building"
 ./build_linux_amd64.sh
 echo "# Cleaning up previous builds"
 rm -rf build
+rm -rf buildtar
 rm *.tar.gz
 rm *.deb
 rm *.rpm
@@ -43,11 +44,15 @@ cp -r share/* build/usr/share/replication-manager/
 rm -rf build/usr/share/replication-manager/opensvc/*.tar.gz
 
 
-for flavor in min osc tst pro arb
+for flavor in min osc tst pro
 do
     echo "# Building packages replication-manager-$flavor"
     cp etc/* build/etc/replication-manager/
-    rm -f build/etc/replication-manager/config.toml.sample.opensvc.*
+    if [ "$flavor" != "pro" ]; then
+      rm -f build/etc/replication-manager/config.toml.sample.opensvc.*
+    else
+      cp -rp test/opensvc build/usr/share/replication-manager/tests
+    done
     cp replication-manager-$flavor build/usr/bin/
     cp service/replication-manager-$flavor.service build/etc/systemd/system/replication-manager.service
     cp service/replication-manager-$flavor.init.el6 build/etc/init.d/replication-manager
@@ -58,12 +63,26 @@ do
 
     echo "# Building tarball replication-manager-$flavor"
     cp etc/* buildtar/etc/
-    rm -f build/etc/config.toml.sample.opensvc.*
+    rm -f buildtar/etc/config.toml.sample.opensvc.*
     cp replication-manager-$flavor buildtar/bin/
     cp service/replication-manager-$flavor.service buildtar/etc/replication-manager.service
     cp service/replication-manager-$flavor.init.el6 buildtar/etc/replication-manager.init
     fpm --package replication-manager-$flavor-$version.tar -C buildtar -s dir -t tar -n replication-manager-$flavor .
     gzip replication-manager-$flavor-$version.tar
+    rm -rf buildtar/usr/bin/replication-manager-$flavor
 done
 
+echo "# Building arbitrator packages"
+rm -rf build/etc
+rm -rf build/usr/share
+cp service/replication-manager-arb.service build/etc/systemd/system
+cp service/replication-manager-arb.init.el6 build/etc/init.d/replication-manager-arb
+cp replication-manager-arb build/usr/bin/
+fpm --rpm-os linux --epoch $epoch --iteration $head -v $version -C build -s dir -t rpm -n replication-manager-arbitrator .
+fpm --package replication-manager-arbitrator-$version-$head.tar -C build -s dir -t tar -n replication-manager-arbitrator .
+gzip replication-manager-arbitrator-$version-$head.tar
+cp service/replication-manager-arb.init.deb7 build/etc/init.d/replication-manager-arbitrator
+fpm --epoch $epoch --iteration $head -v $version -C build -s dir -t deb -n replication-manager-arbitrator .
+rm -f build/usr/bin/replication-manager-arb
 
+echo "# Build complete"
